backend/requirements.txt
Django==4.1.1
python-dotenv==0.19.2
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==5.0.0
djoser ==2.0.1
Pillow == 9.2.0